connected frontend and formatted website
